Changed in the 'Add to date' function the type of the 'year' variable

+++ src/multiCal++.cpp
@@ -7,7 +7,7 @@ float add(float number1, float number2){
 }
 
 
-void add_Days_To_Date(unsigned short day, unsigned short month, int year, int daysToAdd)
+void add_Days_To_Date(unsigned short day, unsigned short month, unsigned short year, int daysToAdd)
 {
 	if(	((month == 1 || month == 3 || month == 5 || month == 7 || month == 8 || month == 10 || month == 12) && day > 0 && day < 32)
 		|| ((month == 4 || month == 6 || month == 9 || month == 11) && day > 0 && day < 31)
@@ -17,7 +17,7 @@ void add_Days_To_Date(unsigned short day, unsigned short month, int year, int da
 	{
 		unsigned short storeDay = day;
 		unsigned short storeMonth = month;
-		int storeYear = year;
+		unsigned short storeYear = year;
 		int storeDaysToAdd = daysToAdd;
